Instructions view in a dedicated file
